# metal : separate scale and mask from QKT in FA kernel