alt https://github.com/ggerganov/llama.cpp/pull/9187

With this change, each thread in the simdgroup accesses the same data (`ss[j*TF + tiisg]`) when applying the scale/mask/logitcap and the softmax after that. This way, no synchronization is necessary

- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [ ] Low
  - [ ] Medium
  - [ ] High