@slaren 2024-08-26
This also works on M3 Max. Though if the data is actually local to the thread, then storing it in a local may be faster than using shared memory.