@ggerganov 2024-08-26
Yup, there is a small improvement by keeping it local:

```bash
make -j && ./scripts/compare-commits.sh \
    e865686c218a5eb2be1a537e3a5e7b9a2acefdde \
    ff23e8e9f09be1eefefd9e1940aa1018854604fe \
    -m ./models/tinyllama-1b/ggml-model-f16.gguf \
    -m ./models/tinyllama-1b/ggml-model-q8_0.gguf \
    -m ./models/tinyllama-1b/ggml-model-q4_0.gguf \
    -m ./models/llama-8b-v3/ggml-model-f16.gguf -r 10 -fa 1 -t 4
```

| CPU   | Model         |   Model Size [GiB] |    Test   |   t/s e865686c |   t/s gg/metal-fix-fa-3 |   Speedup |
|:------|:--------------|-------------------:|-------|---------------:|------------------------:|----------:|
| M2 Ultra | llama 1B F16  |               2.05 | pp512  |        7620.16 |                 7711.19 |      1.01 |
| M2 Ultra | llama 1B F16  |               2.05 | tg128  |         150.25 |                  151.36 |      1.01 |
| M2 Ultra | llama 1B Q4_0 |               0.59 | pp512  |        6965.17 |                 7041.40 |      1.01 |
| M2 Ultra | llama 1B Q4_0 |               0.59 | tg128  |         242.68 |                  243.96 |      1.01 |
| M2 Ultra | llama 1B Q8_0 |               1.09 | pp512  |        6875.29 |                 6924.46 |      1.01 |
| M2 Ultra | llama 1B Q8_0 |               1.09 | tg128  |         207.83 |                  207.43 |      1.00 |
| M2 Ultra | llama 8B F16  |              14.96 | pp512  |        1358.81 |                 1396.95 |      1.03 |
| M2 Ultra | llama 8B F16  |              14.96 | tg128  |          38.84 |                   38.89 |      1.00 |