@slaren 2024-08-26
I tried to test it, but I am afraid that this MBP doesn't handle the summer heat very well, and the results are very inconsistent even with a large number of repetitions. Overall it seems faster, though.

<details>

| CPU   | Model         |   Model Size [GiB] |   Num. of Par. | Test   |   t/s master |   t/s gg/metal-fix-fa-3 |   Speedup |
|:------|:--------------|-------------------:|---------------:|:-------|-------------:|------------------------:|----------:|
|       | llama 1B Q8_0 |               1.09 |     1100048384 | pp512  |      4921.15 |                 4908.74 |      1.00 |
|       | llama 7B F16  |              12.55 |     6738415616 | pp512  |       823.43 |                  861.02 |      1.05 |

| CPU   | Model         |   Model Size [GiB] |   Num. of Par. | Test   |   t/s master |   t/s gg/metal-fix-fa-3 |   Speedup |
|:------|:--------------|-------------------:|---------------:|:-------|-------------:|------------------------:|----------:|
|       | llama 1B Q8_0 |               1.09 |     1100048384 | pp2048 |      4425.88 |                 3649.76 |      0.82 |
|       | llama 7B F16  |              12.55 |     6738415616 | pp2048 |       627.21 |                  452.36 |      0.72 |
|       | llama 7B Q4_0 |               3.56 |     6738415616 | pp2048 |       660.62 |                  375.44 |      0.57 |

| CPU   | Model         |   Model Size [GiB] |   Num. of Par. | Test   |   t/s master |   t/s gg/metal-fix-fa-3 |   Speedup |
|:------|:--------------|-------------------:|---------------:|:-------|-------------:|------------------------:|----------:|
|       | llama 1B Q8_0 |               1.09 |     1100048384 | pp2048 |      2503.88 |                 2598.13 |      1.04 |
|       | llama 7B F16  |              12.55 |     6738415616 | pp2048 |       331.05 |                  372.34 |      1.12 |
|       | llama 7B Q4_0 |               3.56 |     6738415616 | pp2048 |       440.86 |                  473.36 |      1.07 |

| CPU   | Model         |   Model Size [GiB] |   Num. of Par. | Test   |   t/s master |   t/s gg/metal-fix-fa-3 |   Speedup |
|:------|:--------------|-------------------:|---------------:|:-------|-------------:|------------------------:|----------:|
|       | llama 1B Q8_0 |               1.09 |     1100048384 | pp2048 |      2529.39 |                 2985.77 |      1.18 |
|       | llama 7B F16  |              12.55 |     6738415616 | pp2048 |       392.69 |                  426.41 |      1.09 |
|       | llama 7B Q4_0 |               3.56 |     6738415616 | pp2048 |       500.22 |                  511.85 |      1.02 |

| CPU   | Model         |   Model Size [GiB] |   Num. of Par. | Test   |   t/s master |   t/s gg/metal-fix-fa-3 |   Speedup |
|:------|:--------------|-------------------:|---------------:|:-------|-------------:|------------------------:|----------:|
|       | llama 1B Q8_0 |               1.09 |     1100048384 | pp512  |      3095.63 |                 3463.95 |      1.12 |
|       | llama 7B F16  |              12.55 |     6738415616 | pp512  |       425.79 |                  434.49 |      1.02 |
|       | llama 7B Q4_0 |               3.56 |     6738415616 | pp512  |       526.61 |                  523.61 |      0.99 |

</details>